Add files via upload
Initial file (Uncomment part 1 and 2 according to the question for the solutions)

Question: Find the analytical solution of the following two ODEs and plot the result. Use Euler’s method to find the solution ∀ x ∈[0,50] with a step size of 0.1 and 0.05. Compare
the analytical solution with the numerical solution. Use Python
i. dy/dx −5y=−5/2*y3
Find y(0) & use it as Initial condition.
ii. (xy−1)*dx + (x^2−xy)*dy = 0 Find y(0) & use it as Initial condition.
iii. Code and Plot of analytical solution